# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scenic Routes MCP API – orchestrates Google Maps & Shopify for scenic merchant discovery
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RouteWaypoint {
  lat: number;
  lng: number;
  /** @nullable */
  name: string | null;
}

export interface ScenicRoute {
  /** Google Maps encoded polyline string */
  encodedPolyline: string;
  distanceKm: number;
  durationMinutes: number;
  waypoints: RouteWaypoint[];
  /** Human-readable route summary */
  summary: string;
  /** Static Google Maps URL showing the route */
  mapUrl: string;
}

export interface MerchantProduct {
  id: string;
  title: string;
  price: string;
  /** @nullable */
  imageUrl: string | null;
  checkoutUrl: string;
}

export interface Merchant {
  id: string;
  name: string;
  /** winery, bakery, artisan, cafe, etc. */
  type: string;
  lat: number;
  lng: number;
  address: string;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  distanceFromRouteKm: number | null;
  /** @nullable */
  photoUrl: string | null;
  description: string;
  /** @nullable */
  isOpen: boolean | null;
}

export interface MerchantCardRequest {
  merchantId: string;
  merchantName: string;
  merchantType: string;
}

export interface MerchantCard {
  merchant: Merchant;
  products: MerchantProduct[];
  /** One-tap Shopify checkout URL */
  checkoutUrl: string;
  /** Distance in km at which a wearable alert should trigger */
  hapticDistance: number;
}

/**
 * JSON Schema for the tool's input parameters
 */
export type McpToolInputSchema = { [key: string]: unknown };

export interface McpTool {
  name: string;
  description: string;
  /** JSON Schema for the tool's input parameters */
  inputSchema: McpToolInputSchema;
}

export type GetScenicRouteParams = {
  /**
   * Origin address or lat,lng (e.g. "Toronto, ON")
   */
  origin: string;
  /**
   * Destination address or lat,lng (e.g. "Niagara-on-the-Lake, ON")
   */
  destination: string;
};

export type GetMerchantsParams = {
  /**
   * Latitude
   */
  lat: number;
  /**
   * Longitude
   */
  lng: number;
  /**
   * Search radius in meters (default 5000)
   */
  radius?: number;
  /**
   * Merchant type filter (winery, bakery, artisan, all)
   */
  type?: string;
};