# lib/api-zod/src/generated/types/mcpTool.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scenic Routes MCP API – orchestrates Google Maps & Shopify for scenic merchant discovery
 * OpenAPI spec version: 0.1.0
 */
import type { McpToolInputSchema } from "./mcpToolInputSchema";

export interface McpTool {
  name: string;
  description: string;
  /** JSON Schema for the tool's input parameters */
  inputSchema: McpToolInputSchema;
}