# lib/api-zod/src/generated/types/merchant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scenic Routes MCP API – orchestrates Google Maps & Shopify for scenic merchant discovery
 * OpenAPI spec version: 0.1.0
 */

export interface Merchant {
  id: string;
  name: string;
  /** winery, bakery, artisan, cafe, etc. */
  type: string;
  lat: number;
  lng: number;
  address: string;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  distanceFromRouteKm: number | null;
  /** @nullable */
  photoUrl: string | null;
  description: string;
  /** @nullable */
  isOpen: boolean | null;
}