# lib/api-zod/src/generated/types/merchantCard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scenic Routes MCP API – orchestrates Google Maps & Shopify for scenic merchant discovery
 * OpenAPI spec version: 0.1.0
 */
import type { Merchant } from "./merchant";
import type { MerchantProduct } from "./merchantProduct";

export interface MerchantCard {
  merchant: Merchant;
  products: MerchantProduct[];
  /** One-tap Shopify checkout URL */
  checkoutUrl: string;
  /** Distance in km at which a wearable alert should trigger */
  hapticDistance: number;
}