# lib/api-zod/src/generated/types/routeWaypoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scenic Routes MCP API – orchestrates Google Maps & Shopify for scenic merchant discovery
 * OpenAPI spec version: 0.1.0
 */

export interface RouteWaypoint {
  lat: number;
  lng: number;
  /** @nullable */
  name: string | null;
}