# lib/api-zod/src/generated/types/scenicRoute.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scenic Routes MCP API – orchestrates Google Maps & Shopify for scenic merchant discovery
 * OpenAPI spec version: 0.1.0
 */
import type { RouteWaypoint } from "./routeWaypoint";

export interface ScenicRoute {
  /** Google Maps encoded polyline string */
  encodedPolyline: string;
  distanceKm: number;
  durationMinutes: number;
  waypoints: RouteWaypoint[];
  /** Human-readable route summary */
  summary: string;
  /** Static Google Maps URL showing the route */
  mapUrl: string;
}